# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate by identifier + password (api-spec \u00a72.3)",
        "description": "Establish a session: mint an access token + set the HttpOnly refresh\ncookie. Non-enumerating 401 for a wrong identifier OR password; 423 (with\n`retry_after_seconds`) while locked; 403 for inactive accounts (api-spec \u00a72.3).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "Locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotate the refresh token and re-issue an access token (api-spec \u00a72.3)",
        "description": "Cookie-only auth + the custom `X-Refresh: 1` header (CSRF defense, \u00a72.2).\nValidates + ROTATES the refresh session. On ANY failure \u2192 401 refresh_invalid\nAND the cookie is cleared.\n\nNOTE: every failure path RETURNS a `JSONResponse` (rather than raising) so the\nclearing `Set-Cookie` is attached to the response that is actually sent. If we\nraised, FastAPI would route to the global handler and DISCARD this `response`\nobject \u2014 the clearing cookie would never reach the browser (api-spec \u00a72.3\nrequires a refresh failure to clear the cookie).",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "X-Refresh",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Terminate the current session (idempotent, api-spec \u00a72.3)",
        "description": "Authenticated; revoke the refresh session named by the cookie and clear it.\nIdempotent \u2014 an already-logged-out caller still gets 204.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The signed-in user's identity, role and preferences (api-spec \u00a72.3)",
        "description": "Resolve the CurrentUser payload (profile ids + preferences).",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me/password": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Change your own password (api-spec \u00a72.3, FR-AUTH-09)",
        "description": "Self only. `current_password` required UNLESS must_change_password is set.\nNew password is validated against the policy \u2192 422 weak_password. On success,\nall OTHER refresh sessions are revoked (the current one is kept).",
        "operationId": "change_my_password_api_v1_auth_me_password_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/users/{user_id}/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Admin-initiated password reset (api-spec \u00a72.3, FR-SET-04)",
        "description": "Principal/Secretary reset another user. A Secretary may NOT reset a\nPrincipal (403); cannot reset self (409 cannot_reset_self); unknown target\n(404 user_not_found). Returns a generated temporary password ONCE.",
        "operationId": "reset_user_password_api_v1_auth_users__user_id__reset_password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUser"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "AuthTokenResponse",
        "description": "200 body of /auth/login and /auth/refresh."
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "New Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "new_password"
        ],
        "title": "ChangePasswordRequest"
      },
      "CurrentUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "student_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Profile Id"
          },
          "teacher_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Profile Id"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferences"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "must_change_password",
          "is_active",
          "preferences"
        ],
        "title": "CurrentUser",
        "description": "GET /auth/me + login/refresh `user`. Byte-for-byte matches the frontend's\n`CurrentUser` (shared/types/api.ts)."
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Identifier"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "delivery": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery",
            "default": "temp_password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ResetPasswordRequest"
      },
      "ResetPasswordResponse": {
        "properties": {
          "temporary_password": {
            "type": "string",
            "title": "Temporary Password"
          }
        },
        "type": "object",
        "required": [
          "temporary_password"
        ],
        "title": "ResetPasswordResponse",
        "description": "200 body of /auth/users/{id}/reset-password \u2014 returned ONCE."
      },
      "Role": {
        "type": "string",
        "enum": [
          "principal",
          "secretary",
          "teacher",
          "student"
        ],
        "title": "Role"
      },
...
      "UserPreferences": {
        "properties": {
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "date_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Format"
          },
          "default_page_size": {
            "type": "integer",
            "title": "Default Page Size"
          }
        },
        "type": "object",
        "required": [
          "locale",
          "theme",
          "default_page_size"
        ],
        "title": "UserPreferences"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}